Printing access logs to console rather than file
Since this Nginx configuration is meant to be used as a Router running as a container in OpenShift, access logs should be printed to the console log in order to be accessible from the console and collected with Elastic. It will also prevent the local container storage from running full.